+++ src/main/kotlin/com/personio/synthetics/domain/SyntheticTestParameters.kt
@@ -1,0 +1,7 @@
+package com.personio.synthetics.domain
+
+data class SyntheticTestParameters(
+    var message: String,
+    var locations: List<String>,
+    var tags: MutableList<String>
+)