+++ src/main/kotlin/com/personio/synthetics/model/config/Timeframe.kt
@@ -3,17 +3,14 @@ package com.personio.synthetics.model.config
 import java.time.DayOfWeek
 import java.time.LocalTime
 
-class Timeframe private constructor(
-    val from: LocalTime,
-    val to: LocalTime,
-    val days: Set<DayOfWeek>
-) {
+class Timeframe private constructor(val from: LocalTime, val to: LocalTime, vararg val days: DayOfWeek) {
     companion object {
-        operator fun invoke(
-            from: LocalTime,
-            to: LocalTime,
-            day: DayOfWeek,
-            vararg moreDays: DayOfWeek
-        ) = Timeframe(from, to, setOf(day) + moreDays.toSet())
+        operator fun invoke(from: LocalTime, to: LocalTime, vararg days: DayOfWeek): Timeframe {
+            require(from < to) {
+                "From time must be earlier than To time."
+            }
+
+            return Timeframe(from, to, *days)
+        }
     }
 }