+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Dependencies: updates of dependency versions
 
 ## [Unreleased] - yyyy-mm-dd
+### New features & improvements
+- Add support for Multi-Step API Synthetic Tests
 
 ### Breaking changes